// types/storyblok.ts
export interface LinkBlok {
  _uid?: string;
  label?: string;
  path?: string;
  style?:
    | "filled-dark"
    | "filled-light"
    | "outline-dark"
    | "outline-light"
    | "link";
  icon?: "arrow-right" | "arrow-external-link";
  fullWidth?: boolean;
}

export interface MediaBlok {
  filename?: string;
  alt?: string;
  name?: string;
}